requirements.txt
aiogram==3.0.0b7
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.0  # For Excel export
openpyxl==3.1.2  # For Excel export